pyproject.toml
[tool.poetry]
name = "bareasgi"
version = "4.0.0"
description = "A lightweight ASGI framework"
authors = ["Rob Blackbourn <rob.blackbourn@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "bareasgi" }
]
repository = "https://github.com/rob-blackbourn/bareasgi"

[tool.poetry.dependencies]
python = "^3.8"
bareutils = "^4.0.0"
jetblack-asgi-typing = "^0.4.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
pytest = "^5.0"
pytest-asyncio = "^0.10"
pytest-runner = "^5.1"
uvicorn = "^0.15"
hypercorn = "^0.11"
uvloop = "^0.15"
pylint = "^2.8"
mypy = "^0.910"
rope = "^0.14.0"
mkdocs = "^1.2"
mkdocs-material = "^7.2"
jetblack-markdown = "^0.6"
mike = "^1.1.1"
types-setuptools = "^57.4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
